Arrange predicate files
