+++ src/primitives/Primitives/Predicates/PredicateExtensions.Null.cs
@@ -1,0 +1,17 @@
+﻿#nullable enable
+
+using System.Diagnostics.CodeAnalysis;
+
+namespace System
+{
+    partial class PredicateExtensions
+    {
+        public static bool IsNotNull<T>([NotNullWhen(true)][MaybeNullWhen(false)] this T value)
+            =>
+            value is not null;
+
+        public static bool IsNull<T>([NotNullWhen(false)][MaybeNullWhen(true)] this T value)
+            =>
+            value is null;
+    }
+}
