+++ src/primitives/Primitives/Predicates/PredicateExtensions.NullOrWhiteSpace.cs
@@ -1,0 +1,13 @@
+﻿#nullable enable
+
+using System.Diagnostics.CodeAnalysis;
+
+namespace System
+{
+    partial class PredicateExtensions
+    {
+        public static bool IsNullOrWhiteSpace([NotNullWhen(false)] this string? value)
+            =>
+            string.IsNullOrWhiteSpace(value);
+    }
+}
