+++ src/primitives/Primitives/Predicates/Predicates.NullOrEmpty.cs
@@ -1,0 +1,13 @@
+﻿#nullable enable
+
+using System.Diagnostics.CodeAnalysis;
+
+namespace System
+{
+    partial class Predicates
+    {
+        public static bool IsNullOrEmpty([NotNullWhen(false)] string? value)
+            =>
+            string.IsNullOrEmpty(value);
+    }
+}
