+++ src/primitives/Primitives/Predicates/Predicates.NullOrWhiteSpace.cs
@@ -1,0 +1,13 @@
+﻿#nullable enable
+
+using System.Diagnostics.CodeAnalysis;
+
+namespace System
+{
+    partial class Predicates
+    {
+        public static bool IsNullOrWhiteSpace([NotNullWhen(false)] string? value)
+            =>
+            string.IsNullOrWhiteSpace(value);
+    }
+}
